// src/types/buttons.ts
export enum Buttons {
  addHead = "addHead",
  addTail = "addTail",
  deleteHead = "deleteHead",
  deleteTail = "deleteTail",
  addByIndex = "addByIndex",
  deleteByIndex = "deleteByIndex",
  clear = "clear",
  reverse = "reverse",
  fibonacci = "fibonacci",
  sortDescending = "sortDescending",
  sortAscending = "sortAscending"
}